+++ plutus-extra/test/Suites/PlutusTx/Data/Extra.hs
@@ -17,35 +17,27 @@ import Test.QuickCheck (allProperties)
 
 --------------------------------------------------------------------------------
 
-import Ledger.Typed.Scripts (DatumType, RedeemerType, ValidatorTypes)
+import PlutusTx.Data.Extra (fromDatum, fromRedeemer, toDatum, toRedeemer)
 import PlutusTx.IsData.Class (FromData, ToData)
 
 --------------------------------------------------------------------------------
-import PlutusTx.Data.Extra (fromDatum, fromRedeemer, toDatum, toRedeemer)
-
---------------------------------------------------------------------------------
-
-data Test a
-instance ValidatorTypes (Test a) where
-  type RedeemerType (Test a) = a
-  type DatumType (Test a) = a
 
 prop_ToFromDatumRoundTrip ::
   forall (a :: Type).
-  (ToData (DatumType (Test a)), FromData (DatumType (Test a)), Eq (DatumType (Test a))) =>
-  DatumType (Test a) ->
+  (ToData a, FromData a, Eq a) =>
+  a ->
   Bool
-prop_ToFromDatumRoundTrip x = fromDatum @(Test a) (toDatum @(Test a) x) == Just x
+prop_ToFromDatumRoundTrip x = fromDatum @a (toDatum @a x) == Just x
 
 prop_ToFromRedeemerRoundTrip ::
   forall (a :: Type).
-  (ToData (RedeemerType (Test a)), FromData (RedeemerType (Test a)), Eq (RedeemerType (Test a))) =>
-  RedeemerType (Test a) ->
+  (ToData a, FromData a, Eq a) =>
+  a ->
   Bool
-prop_ToFromRedeemerRoundTrip x = fromRedeemer @(Test a) (toRedeemer @(Test a) x) == Just x
+prop_ToFromRedeemerRoundTrip x = fromRedeemer @a (toRedeemer @a x) == Just x
 
 -- Don't ask me.
 pure []
 
 tests :: [TestTree]
-tests = [testProperties "Date.Extra properties" $allProperties]
+tests = [testProperties "Data.Extra properties" $allProperties]